requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
Django==2.2.7
django-bootstrap-form==3.4
djangorestframework==3.10.3
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
matplotlib==3.1.2
numpy==1.17.4
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
pandas==0.25.3
Pillow==6.2.1
protobuf==3.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyparsing==2.4.5
PyPDF2==1.26.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
urllib3==1.25.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
xhtml2pdf==0.2.3
